fix: 리뷰 반영

+++ src/main/java/com/example/demo/domain/artworkcommunication/application/result/ArtworkQuestionReplyLikeResult.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record ArtworkQuestionReplyLikeResult(
     Long questionReplyId,
     Boolean liked,
-    Integer likeCount,
+    Long likeCount,
     LocalDateTime createdAt,
     LocalDateTime deletedAt) {}
