+++ src/main/java/com/example/demo/domain/artworkcommunication/presentation/response/ArtworkQuestionReplyLikeResponse.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record ArtworkQuestionReplyLikeResponse(
     Long questionReplyId,
     Boolean liked,
-    Integer likeCount,
+    Long likeCount,
     LocalDateTime createdAt,
     LocalDateTime deletedAt) {}
